# go/model_article.go
/*
 * Minimalist Blog
 *
 * This is a sample Blog server. 
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

 package swagger

 type Article struct {
 
	 Name string `json:"name,omitempty"`
 
	 Tag string `json:"tag,omitempty"`
 
	 Author string `json:"author,omitempty"`
 
	 Content string `json:"content,omitempty"`
 
	 Review string `json:"review,omitempty"`
 }
 