# go/model_user.go
/*
 * Minimalist Blog
 *
 * This is a sample Blog server. 
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

 package swagger

 type User struct {
 
	 Username string `json:"username,omitempty"`
 
	 Password string `json:"password,omitempty"`
 }
 