# go/routers.go
/*
 * Minimalist Blog
 *
 * This is a sample Blog server.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"io/ioutil"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

type Route struct {
	Name        string
	Method      string
	Pattern     string
	HandlerFunc http.HandlerFunc
}

type Routes []Route

func NewRouter() *mux.Router {
	router := mux.NewRouter().StrictSlash(true)
	for _, route := range routes {
		var handler http.Handler
		handler = route.HandlerFunc
		handler = Logger(handler, route.Name)

		router.
			Methods(route.Method).
			Path(route.Pattern).
			Name(route.Name).
			Handler(handler)
	}

	return router
}

func Index(w http.ResponseWriter, r *http.Request) {
	str, _ := ioutil.ReadFile("./go/api.json")
	w.WriteHeader(http.StatusOK)
	w.Header().Set("Content-Type", "application/json")
	w.Write(str)
}

var routes = Routes{
	Route{
		"Index",
		"GET",
		"/v3/",
		Index,
	},

	Route{
		"GetArticlesByAuthor",
		strings.ToUpper("Get"),
		"/v3/articles/GetArticlesByAuthor",
		GetArticlesByAuthor,
	},

	Route{
		"GetArticlesByName",
		strings.ToUpper("Get"),
		"/v3/articles/GetArticlesByName",
		GetArticlesByName,
	},

	Route{
		"GetArticlesByTag",
		strings.ToUpper("Get"),
		"/v3/articles/GetArticlesByTag",
		GetArticlesByTag,
	},

	Route{
		"LogOut",
		strings.ToUpper("Get"),
		"/v3/user/logout",
		LogOut,
	},

	Route{
		"SignIn",
		strings.ToUpper("Get"),
		"/v3/user/sign_in",
		SignIn,
	},

	Route{
		"SignUp",
		strings.ToUpper("Post"),
		"/v3/user/sign_up",
		SignUp,
	},
}